feat(crd): refactoring CRD
Changed the group name from openebs.io to zfs.openebs.io.
