+++ deploy/zfs-operator.yaml
@@ -10,9 +10,9 @@ metadata:
 apiVersion: apiextensions.k8s.io/v1beta1
 kind: CustomResourceDefinition
 metadata:
-  name: zfsvolumes.openebs.io
+  name: zfsvolumes.zfs.openebs.io
 spec:
-  group: openebs.io
+  group: zfs.openebs.io
   version: v1alpha1
   scope: Namespaced
   names:
@@ -57,9 +57,9 @@ spec:
 apiVersion: apiextensions.k8s.io/v1beta1
 kind: CustomResourceDefinition
 metadata:
-  name: zfssnapshots.openebs.io
+  name: zfssnapshots.zfs.openebs.io
 spec:
-  group: openebs.io
+  group: zfs.openebs.io
   version: v1alpha1
   scope: Namespaced
   names:
