+++ pkg/builder/volbuilder/volume.go
@@ -15,7 +15,7 @@
 package volbuilder
 
 import (
-	apis "github.com/openebs/zfs-localpv/pkg/apis/openebs.io/core/v1alpha1"
+	apis "github.com/openebs/zfs-localpv/pkg/apis/openebs.io/zfs/v1alpha1"
 )
 
 // ZFSVolume is a wrapper over
